- make xmview compile cleanly on x86_64 (this does not mean that it will work   correctly - it still has to be tested):   * int changed to long where needed and L appended to numeric constants where     needed   * getregs/setregs converted from #define to inline and removed references to     "regs_modified" from these and other similar functions (get/set args, scno     etc.)   * removed duplicated definition of VIEWOS_FRAME_SIZE   * fixed mmap management (x86_64 doesn't have mmap2): on x86_64 the offset in     bytes is still rounded up to a multiple of the page size but mmap is     called instead of mmap2 with offset expressed in bytes   * fixed __NR_nice (not-existent on x86_64, implemented using setpriority)   * fixed __NR_doesnotexist logic for modules: __NR_xxx64 constants for the     64-bit version of some system calls must be defined equal to __NR_xxx (and     not __NR_doesnotexist) as the 64-bit version on x86_64 has no "64" suffix. The modules     interfaces says that the syscalls to be defined are the xxx64 ones also on     64 bit platforms, so lots of modules code expects to have a valid __NR_xxx64 constant defined. - umNULL defined in defs.h instead of every um_*.c - added some clarifying comments about scmap - fixed return types for send/recv/sendto/recvfrom/... socket functions (was   int, should be ssize_t)

+++ xmview-os/include/umnet.h
@@ -50,14 +50,14 @@ struct umnet_operations {
 	int (*accept) (int, struct sockaddr *, socklen_t *);
 	int (*getsockname) (int, struct sockaddr *, socklen_t *);
 	int (*getpeername) (int, struct sockaddr *, socklen_t *);
-	int (*send) (int, const void *, size_t, int) ;
-	int (*recv) (int, const void *, size_t, int);
-	int (*sendto) (int, const void *, size_t, int, const struct sockaddr *, socklen_t);
-	int (*recvfrom) (int, void *, size_t, int, struct sockaddr *, socklen_t *);
+	ssize_t (*send) (int, const void *, size_t, int) ;
+	ssize_t (*recv) (int, const void *, size_t, int);
+	ssize_t (*sendto) (int, const void *, size_t, int, const struct sockaddr *, socklen_t);
+	ssize_t (*recvfrom) (int, void *, size_t, int, struct sockaddr *, socklen_t *);
 	int (*setsockopt) (int, int, int, const void *, socklen_t);
 	int (*getsockopt) (int, int, int, void *, socklen_t *);
-	int (*read) (int, void *, size_t);
-	int (*write) (int, const void *, size_t);
+	ssize_t (*read) (int, void *, size_t);
+	ssize_t (*write) (int, const void *, size_t);
 	int (*ioctl) (int, int, void *);
 	int (*close) (int);
